device-setup/service/requirements.txt
CoAPthon3
requests
mysql-connector-python
